[FIX] stock_batch_picking: add tests
