+++ stock_batch_picking/tests/__init__.py
@@ -1,1 +1,2 @@
 from . import test_batch
+from . import test_print_batch
